API: Use SettingsProvider on GET settings handler (#34632)
