+++ pkg/setting/provider.go
@@ -2,6 +2,7 @@ package setting
 
 import (
 	"errors"
+	"regexp"
 	"strings"
 	"time"
 
@@ -32,7 +33,13 @@ func (v ValidationError) Error() string {
 // Provider is a settings provider abstraction
 // with thread-safety and runtime updates.
 type Provider interface {
-	// Update
+	// Current returns a SettingsBag with a static copy of
+	// the current configured pairs of key/values for each
+	// configuration section.
+	Current() SettingsBag
+	// Update receives a SettingsBag with the pairs of key/values
+	// to be updated per section and a SettingsRemovals with the
+	// section keys to be removed.
 	Update(updates SettingsBag, removals SettingsRemovals) error
 	// KeyValue returns a key-value abstraction
 	// for the given pair of section and key.
@@ -92,6 +99,37 @@ type OSSImpl struct {
 
 func (o OSSImpl) Init() error {
 	return nil
+}
+
+func (o OSSImpl) Current() SettingsBag {
+	settingsCopy := make(SettingsBag)
+
+	for _, section := range o.Cfg.Raw.Sections() {
+		settingsCopy[section.Name()] = make(map[string]string)
+
+		for _, key := range section.Keys() {
+			keyName := key.Name()
+			value := key.Value()
+
+			if strings.Contains(keyName, "secret") || strings.Contains(keyName, "password") || (strings.Contains(keyName, "provider_config")) {
+				value = "************"
+			}
+
+			if strings.Contains(keyName, "url") {
+				var rgx = regexp.MustCompile(`.*:\/\/([^:]*):([^@]*)@.*?$`)
+				var subs = rgx.FindAllSubmatch([]byte(value), -1)
+
+				if subs != nil && len(subs[0]) == 3 {
+					value = strings.Replace(value, string(subs[0][1]), "******", 1)
+					value = strings.Replace(value, string(subs[0][2]), "******", 1)
+				}
+			}
+
+			settingsCopy[section.Name()][keyName] = value
+		}
+	}
+
+	return settingsCopy
 }
 
 func (OSSImpl) Update(SettingsBag, SettingsRemovals) error {
